Add files via upload
I apologize for the very rough presentation in the attached documents.

In the file "Ulam.nb", there is a sequence of small functions (e.g., ulam[n_], ulamslist[n_], ulamnumber[n_], etc.) that calculate a variety of properties related to the 3N+1 problem. For example, ulam[n_] calculates the next value in the 3N+1 sequence given an input number n. The function ulamslist[n_] generates the 3N+1 sequence for a number n, and the function ulamnumber[n_] calculates the number of terms in the list. Here, I am not attempting to solve the 3N+1 problem, the goal is simply to implement an observation. It has been observed that if a number is a power of 2, then the 3N+1 sequence will be a decreasing sequence of numbers ending in 4-2-1, where each term is one-half the preceding term. One generalization of this observation is to consider the value of a number reduced by a modulus of a power of two. Here, I found that eight is useful. If a number in the sequence is congruent to 0, 2, 4, or 6 (mod 8), the number immediately undergoes a reduction (by at least one to three steps of division by 2). In these cases, the number is reduced. When a number n is congruent to 1 or 5 (mod 8), then we have a net decrease. This happens for n congruent to (1 mod 8) since multiplying by 3 and adding 1 leads to a number congruent to (1*3 + 1) = 4 (mod 8). The number increases roughly by a factor of 3 and the result is reduced by a factor of (at least) four. The overall change is approximately three-fourths, a net reduction. A similar decrease occurs for a number congruent to 5 (mod 8). A number increases only if the number is congruent to 3 or 7 (modulo 8). It is well-known that the number 27 has relatively large values contained in its 3N+1 sequence. It is then likely that a number of terms in its 3N+1 sequence are congruent to 3 or 7 (mod 8). The graphical representation given by the function ulamgraph[n_] shows the value of each successive term reduced modulo 8. Numbers that are a multiple of 8 lie on the positive horizontal axis. Numbers that are 1 (mod 8) lie on the line y=x in the first quadrant. Every additional angle of one-fourth pi radians leads to another of the 8 possible values (modulo 8). In particular, the numbers that are congruent to 7 (mod 8) lie along the line y = -x in the fourth quadrant. In the graph for ulamgraph[27], one can see that there are many values that are congruent to 7 (mod 8). Another approach to the conjecture is to determine the expectation that the numbers in a 3N+1 sequence are congruent to 3 or 7 (modulo 8) since these are the only values that cause an increase in the value of the 3N+1 sequence.